+++ 02-dtmoney/src/App.tsx
@@ -4,6 +4,7 @@ import { GlobalStyle } from "./styles/global"; // Importando o estilo global
 import Modal from 'react-modal';
 import { useState } from "react";
 import { NewTransactionModal } from "./components/NewTransactionModal";
+import { TransactionsContext } from "./TransactionsContext";
 
 Modal.setAppElement('#root'); // Configurando o modal para ser aberto dentro do elemento root
 
@@ -17,11 +18,11 @@ export function App() { //permite que o componente seja exportado com o mesmo no
         setIsNewTransactionModalOpen(false);
     }
 	return (
-		<> {/* O React precisa de um elemento raiz, no caso, o elemento <> </>  que é uma tag fragment */}
+		<TransactionsContext.Provider value={[]}> {/* É um objeto e acessamos a propriedade provider, value é obrigatório e informa o valor atual do contexto*/}
 			<Header onOpenNewTransactionModal={handleOpenNewTransactionModal}/> {/*Passando a função para o componente Header chamar o modal */ }
 			<Dashboard/>
 			<NewTransactionModal isOpen={isNewTransactionModalOpen} onClose={handleCloseNewTransactionModal} />			
 			<GlobalStyle/> {/* Usando o estilo global */}
-		</>
+		</TransactionsContext.Provider>
 	);
 }