+++ 02-dtmoney/src/components/Summary/index.tsx
@@ -2,10 +2,15 @@ import { Container } from "./styles";
 import incomeImg from "../../assets/income.svg";
 import outcomeImg from "../../assets/outcome.svg";
 import totalImg from "../../assets/total.svg";
+import { useContext} from "react";
+import { TransactionsContext } from "../../TransactionsContext";
 
 export function Summary() {
+	const data = useContext(TransactionsContext);
+
     return (
 		<Container>
+
 			<div>
 				<header> {/* Cabeçalho da box */}
 					<p>Entradas</p>